Refactor for usage as library
